Plan-23: split platform_esp32.cpp by subsystem
platform_esp32.cpp goes from 1281 lines to 700 (core: system primitives,
network, sockets, mDNS). Three subsystems move to siblings under the
same dir, each self-contained (own private state, only depends on the
public platform.h surface):

  src/platform/esp32/platform_esp32_fs.cpp        (177 lines)
    LittleFS partition mount + fs* / filesystem* API.
  src/platform/esp32/platform_esp32_ota.cpp       (170 lines)
    esp_https_ota fetch-and-flash task. http_fetch_to_ota entry.
  src/platform/esp32/platform_esp32_improv.cpp    (311 lines)
    UART0 listener, RPC dispatch, credential pump.
    Includes the MM_NO_WIFI link-parity stub.

Network stayed in platform_esp32.cpp — Eth + WiFi + sockets + mDNS
share file-scope state (event handlers, netif pointers, init-done
flags). Splitting them would require an internal header with `extern`
declarations or a singleton refactor. Not in plan-23's scope; carved
out for a future plan if it earns its keep.

Desktop (platform_desktop.cpp) untouched — the OTA / Improv / FS
sections there are 6-line stubs each; per-subsystem files would be
all overhead, no payoff. Asymmetry is intentional.

CMakeLists.txt adds the three new .cpp files to the main component's
SRCS list. No build-system changes beyond that.

Pre-commit gates
----------------
- Desktop build clean.
- Unit tests 172/172, scenarios 8/8.
- All 4 ESP32 boards build:
    esp32         1,260,897 B  (was 1,258,553 — +2.3 KB)
    esp32-eth       906,761 B  (was   904,544 — +2.2 KB)
    esp32-eth-wifi 1,305,573 B  (was 1,303,121 — +2.5 KB)
    esp32s3-n16r8 1,269,521 B  (was 1,267,232 — +2.3 KB)
  ~2.3 KB growth per board from TU-boundary overhead (lost cross-TU
  inlining + per-TU namespace setup). Negligible on a 1.3 MB binary.
- Olimex flashed esp32-eth-wifi, Improv probe responds correctly
  with device info + state — confirms the split didn't break the
  runtime side.
